+++ CHANGELOG.md
@@ -1,3 +1,27 @@
-##v.0001
-    CREATE BOTTOM TAB NAVIGATION
-    EMPTY SCREENS 
+## [v.0000] - 2020-02-23
+
+### Changes
+
+    - 
+
+### Added
+
+    - Initial app config
+
+
+
+## [v.0000] - 2020-02-25
+
+### Changes
+
+    - App.tsx
+
+### Added
+
+    - react-navigation
+    - vector-icons
+    - Screen:
+        Home
+        List
+        Info
+    -styled-components/native